e2e: extract shared futureDate helper & add brand link target size test
- Extract futureDate() into e2e/helpers.ts (DRY)
- a11y.spec: import helper, add brand link 24px target size assertion
- happy-path.spec: import helper, remove inline futureDate()
